Add Product archiving, Not found Page

+++ src/app/components/products/product-preview/product-preview.component.ts
@@ -74,11 +74,13 @@ export class ProductPreviewComponent {
   }
 
   addToCart() {
+    if (this.canSetPic) return;
     this.cartService.addToCart(this.product);
     this.router.navigate(['/panier']);
   }
 
   removeFromCart() {
+    if (this.canSetPic) return;
     this.cartService.removeFromCart(this.product);
   }
 }
